# Use faster foreach_set for keyframe interpolation in 2.90+