The faster foreach_set method now works for setting the keyframe interpolation type in Blender 2.90.

This PR uses the faster method if a supported Blender version is detected. In 2.90 and above, this makes the import always as fast as it currently is with the Bezier interpolation type. Performance in 2.8x is unchanged.

Some performance tests on a 57 MB .agr and Blender 2.90:

| Version | Bezier interpolation import time | Constant interpolation import time |
| -------   | ---------------------------------    | --------------------------------------  |
| Current | 172,8565 s                                    | 482,3850 s                                        |
| This PR | 171,7507 s                                     | 172,2544 s                                        |